pyproject.toml
[project]
name = "rhythmix-model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kagglehub>=0.3.12",
    "langchain-community>=0.3.22",
    "langchain-openai>=0.3.14",
    "langchain-qdrant>=0.2.0",
    "langchain>=0.3.24",
    "langgraph>=0.3.34",
    "pandas>=2.2.3",
    "pre-commit>=4.2.0",
    "python-dotenv>=1.1.0",
    "qdrant-client[fastembed]>=1.14.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
